3aula_openGl_Trianglo/requirements.txt
pygame
pyOpenGL
numpy